Add additional dynamic check if app is sandboxed
This adds a less heuristic check whether an application is sandboxed.
Our current dynamic check could fail, since it is based on some
initialization artefacts. The added check in contrast asks the OS
whether the application is sandboxed. To avoid regressions both checks
are added and the second can be used to validate the results of the
first. The caveat is that it now takes longer to process applications.