Add driver script for analysing all applications
The script iterates over a given folder and runs the appnalyser script
for each application found. The results are written into a given output
folder.

Before running this it is recommended to mount your drive containing the
applications as read-only and to make an APFS snapshot of your test
machine with `tmutil snapshot`.