Fix invalid sandbox status for non-existing processes
When calling `sandbox_check(pid, NULL, 0)` with a `pid` of a process that
was already terminated, `1` will always be returned as a status.
This is misleading, as it indicates that the process is sandboxed,
even though there is no process at all. Therefore, we now check if the
`pid` exists and return an appropriate error instead.

It looks like the Python interpreter keeps a subprocesses PID active
even though the process has terminated, e.g., because it could not be
started. Therefore, we need to query the status of the subprocess
before starting the `sandbox_status` tool.