Fix app container heuristics
The bundle identifier is normalized by the sandbox runtime so that
it is a valid UTI as defined in the UTI documentation [1]. This
caused no problem for MAS apps, since bundle identfiers are managed
by Apple. However, if the application is not distributed through the
MAS, the bundle identifier is an arbitrary developer-chosen string.
This means that such applications might contain invalid characters,
leading to the sandbox container not being found which in turn leads
to invalid detection of dynamic sandboxing.

All invalid characters are now replaced with a '-' (minus). This was
seen with bundle identifier 'uk.co.insili.Markdown+', where the
container path was '~/Library/Containers/uk.co.insili.Markdown-'.

[1] https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/understanding_utis/understand_utis_conc/understand_utis_conc.html#//apple_ref/doc/uid/TP40001319-CH202-CHDHIJDE