Strip extension of application directories
MacOS handles directories ending with .app as applications, even if the
contents of the directory is something completely different. Spotlight
will still include these directories when trying to launch applications,
which will fail of curse. Finder shows the directories as applications
as well, which means that you have to right-click and show contents for
navigation, which is cumbersome. In order to avoid this, we strip the
.app suffix now.